MGMT-9175: Use an interface with a Mac address (#3880)
https://issues.redhat.com/browse/MGMT-9175
Part of epic https://issues.redhat.com/browse/MGMT-10087
The assisted installer agent has changes to
pass all interfaces to the assisted service.
PR for the agent changes:
https://github.com/openshift/assisted-installer-agent/pull/360
Originally, the assisted service selected the first interface
in the list, which might not have a MAC address.
This change now looks for an interface with a MAC
address and uses it for the install config.