MGMT-10713: Soften IP-stack checks for host_network propagation
This commit softens the checks for IP stack when `cluster.HostNetworks`
are propagated from the hosts' inventory. Currently disabled
`cfg.IPv6Support` implies that all the IPv6 subnets will be removed and as
a result dual-stack hosts will create a cluster with only IPv4 host
networks.

This issue is only affecting clusters created using the UI. This is
because UI propagates its forms with `cluster.HostNetworks` property
whereas users of the API are requested to manually provide
`cluster.MachineNetworks`.

This is not a desired behaviour as it prevents from forming a correct
host networks. With this change, if host contains both IPv4 and IPv6
subnets in its inventory, we will form host networks from both stacks
even if `cfg.IPv6Support` is False.

This goes in line with our current approach that `cfg.IPv6Support` refers
to support of IPv6-only clusters and should not affect dual-stack
clusters. The former is because we don't want to allow IPv6-only due to
the fact that api.openshift.com does not have IPv6 endpoint. For
dual-stack clusters this is not an issue as those clusters can connect
using a regular IPv4 endpoint.

With this change the propagation will look as follows

* disabled `cfg.IPv6Support`
  * dual-stack host - HostNetworks with both stacks
  * IPv4 host - HostNetworks with IPv4
  * IPv6 host - empty HostNetworks
* enabled `cfg.IPv6Support`
  * dual-stack host - HostNetworks with both stacks
  * IPv4 host - HostNetworks with IPv4
  * IPv6 host - HostNetworks with IPv6

Please note that this change will not transparently allow for creation
of IPv6-only clusters if the hosts are dual-stack. This is because the
service implements a separate set of checks so that when
`cfg.IPv6Support` is False it is forbidden to select an IPv6 subnet as
any of the networks. If such an attempt is detected, the service will
return the error message `IPv6 is not supported in this setup`.

Closes: MGMT-10713