MGMT-8435: Fix V2UpdateCluster response to include hosts info
A call for V2UpdateCluster returns an updated cluster object
in the payload, which did not include the whole hosts' information.

This change adds a call to customizeHost() that will add this
information to the returned payload.

One use case where this is needed is an update UI does for VIP DHCP
allocation, to which it needs to retrieve the host requested name.
The UI currently worked around the issue by calling v2GetCluster,
including the above-mentioned info.