MGMT-10569: Remove cluster ignition_config_overrides column (#3914)
* MGMT-10569: Remove unused cluster ignition_config_overrides

This was used for discovery ignition overrides before infraEnv existed.
Now it's not used and should be removed to avoid confusion.

* MGMT-10569: Correct make target for generating migrations

This is used to generate a single new migration so the target name
should be singular. This also ensures the target agrees with what is in
the developer documentation.

* MGMT-10569: Test the correct column for override col type change

Previously this was setting a value in the ignition config override
column but the migration was changing the install config override. The
value should have been set in the column that was actually being
changed.

* MGMT-10569: Remove populate infra env migration

This should not be needed in any environment at this point.
Keeping it would mean trying to maintain this migration without using
the cluster model and that's not worth the effort when this is never
expected to be needed again.

* MGMT-10569: Add migration to remove ignition_config_override from clusters table

Automigrate doesn't remove columns so we need to do this manually
It's required to `Exec` here because the gormigrate `DropColumn`
function requires that the model contain the field being dropped.

Additionally it doesn't support the `IF EXISTS` portion so it would
require more logic to handle this migration correctly.