MGMT-9259: Enhance the schedulable masters feature (simplified)
The current API always returns the value of the 'schedulable_masters'
property saved in the database at cluster creation time (defaults to
"false" if omitted). However the UI needs the API to return the default
correct value depending on the number of hosts in the cluster.

The proposed changes return the expected default values in case the
'schedulable_masters' property was not actively set.