MGMT-10780: Explicitly pass OCP version to Hypershift
Hypershift selects an OCP version for the workload based on the value of
OPENSHIFT_INSTALL_RELEASE_IMAGE env variable. In dev-scripts world the
same variable is used to define what is the OCP version of the Hub
cluster. As those two are not always desired to be the same, we are
leveraging the ASSISTED_OPENSHIFT_INSTALL_RELEASE_IMAGE to define
version of Hypershift's workload cluster.

The variable "rename" happens just before we start using Hypershift
commands so that no other installation flow is affected.

Contributes-to: MGMT-10780