MGMT-10678: Add `compatible-agent` validation
This patch adds a new `comptible-agent` validatio that checks that the
agent image expected by the service and the agent image actually used by
the host are compatible.

As neither the service nor the agent use semantic versioning for this
image, the compatibility check just compares the value of the
`AGENT_DOCKER_IMAGE` environment variable of the service with the
`discovery_agent_version` column of the `hosts` table.

This is part of the agent upgrade feature described
[here](https://github.com/openshift/assisted-service/pull/3928).

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>