MGMT-10780: Copy CAPI deploy scripts to ocm-2.5 branch (#3940)
* MGMT-9227: Create deploy_capi_cluster script for CAPI CI (#3690)

This PR creates a script responsible for deploying a cluster using CAPI
in the setup with Infrastructure Operator. It creates a required
InfraEnv, leverages BareMetalHost created by dev-scripts and later
installs HyperShift to install a cluster.

It is meant to be used by the e2e-metal-assisted-operator-capi job.

* NO-ISSUE: Use official Hypershift image in E2E test (#3735)

As the fix that was supposed to fix the issue
https://bugzilla.redhat.com/show_bug.cgi?id=2079177 has been merged
already, we should be safe to use the hypershift:latest instead our
custom build of the image.