MGMT-10666 - Single Node expansion feature support flag (#3886)
Starting OCP 4.11, adding workers to single node clusters is supported.

Currently, our UI shows a warning for all SNO clusters that says that
their expansion is not supported.

The UI needs to know when it should show the warning about users not
being able to expand a SNO cluster, and when it shouldn't (>=4.11).

To allow the UI to do that, the backend should report a feature support
flag for the 4.11 version, and this commit does that.