MGMT-10311: set user_managed_networking to true with sno in ci
As a preparation for the main code where a webhook is validating that
with SNO user_managed_networking is set to true or not set at all.

I am fixing the CI deploy script for ZTP and CAPI so the value will
not be set to false in case SNO deployment is requested.

Currently, this user_managed_networking configuration is simply ignored
if not done correctly, but going forward, this kind of CRD will be rejected.