NO-ISSUE: Move host validation models to `internal/models`
Currently the types used to represent the host validations are part of
the `internal/host` package. But those types will be needed also in the
`internal/host/hostcommands` package in order to check the result of the
`compatible-agent` validation that will be introduced as part of
MGMT-10678. It isn't possible to just use the existing types because
that would introduce a circular dependency between the `internal/host`
and `internal/host/hostcommands` packages. To avoid that this patch
moves the types to a new `internal/models` package.

Note that this is a very large patch in terms of the lines that it
modifies, but most of then are just to introduce the new
`internalmodels` package selector in the places where the relocated
types are currently being used.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>