MGMT-9710: Provide extended networking defaults in default-config API (#3586) (#3946)
This PR extends the default-config API to contain more defaults
regarding network configuration. This is driven by the fact that
Assisted UI allows now to create single- as well as dual-stack clusters
with multiple combinations of IPv4 and IPv6, therefore it's useful to
have an endpoint that will provide suggested values.

Closes: MGMT-9710

Co-authored-by: Mateusz Kowalski <mko@redhat.com>