MGMT-10848: Fix double agent unbind when deregistering a cluster
When a cluster is removed in the kubeAPI the clusterdeployments
controller removes all the cluster references from all the agents. This
should cause the agent controller to unbind each of those agents.

The clusterdeployment controller also calls `DeregisterClusterInternal`
which also unbinds the hosts from the cluster.

This isn't a problem currently because both operations should be doing
the same thing, but it will become a problem when hosts should be moved
back to the infraEnv on unbind only if they are not associated with an
BMH, and the kubeAPI is being used. In this case the agent controller
will know if the host should be moved to the infraEnv or not, but the
unbind call from `DeregisterClusterInternal` would have to assume that
the hosts should _not_ be moved back to the infraEnv as it would also be
used in the SaaS REST API.

This difference in implementation would cause a race and needs to be
resolved.

This PR moves all the SaaS specific logic out of
`DeregisterClusterInternal` and into `V2DeregisterCluster`. This will
ensure that this logic is still run when it is needed, but avoid the
race with the controllers when using the kubeAPI.

https://issues.redhat.com/browse/MGMT-10848