[release-ocm-2.4] BUG 2097820: CI - Retry cluster-info readiness  (#3967)
* BUG-2097700: CI - Retry cluster-info readiness

Several[1] jobs are failing for this, and it looks like
the cluster was just not ready on time. This is an attempt
to retry cluster-info several times before we fail the job.

[1] one example: periodic-ci-openshift-assisted-service-release-ocm-2.5-subsystem-kubeapi-aws-periodic

* BUG-2097700: Call cluster-info with a namespace

In some cases, the user does not own the cluster and thus cannot
run cluster-info without the scope of a namespace.

Co-authored-by: Nir Magnezi <nmagnezi@redhat.com>