MGMT-10848: Don't require a host to have a cluster to get it by KubeKey
Originally this method just accessed the host, but the cluster
information was added to handle progress reporting.

This left the agent controller with no way to query the database for
hosts who are linked to a possibly deleted cluster.

This can happen when a ClusterDeployment is deleted while it still
had bound agents. The agents will be unbound asynchronously but were
failing because this method required the referenced cluster to still
exist.