MGMT-8089: Force schedulableMasters
Create a new cluster property called `schedulableMastersForcedTrue` that
overrides `scheduleMasters` regardless its value when the number of
hosts (5) doesn't match the requirement for `scheduleMasters` to be
disabled.

This new property can be fetched from the UI and react accorglingly
depending if the value for `scheduleMasters` can be set by the user or
not (e.g.: if `schedulableMastersForcedTrue == true` then disable the
toggle in the UI).