Add subcommands to ephemeral installer client
Change the command to register the Cluster and InfraEnv from:

    /agent-based-installer-register-cluster-and-infraenv

to:

    /usr/local/bin/agent-installer-client register

This allows us to add further subcommands to the same binary in future.

A symlink is provided for backward compatibility with the existing
agent-based installer ISOs. When invoked using the old name, the command
will retain the existing behaviour. This backward compatibility can be
removed once the agent-based installer is using the subcommand syntax.