MGMT-10790: Always use APIVipDNSName to obtain API address
Presently we fetch the API address for ignition download and validation
using two independent pieces of code, these also prioritise ApiVIP and
ApiVIPDNSName in different ways, which leads to incorrect validation.

This commit addresses that by extracting this logic into a function that
may be called from both locations.

During the development of this feature, it was decided that the only
valid address during installation is cluster.APIVipDNSName and so this
function has been created to only return that.

At least for future modification, this logic is in a single location.

Additionally, the validation of whether or not an ignition is downloadable for Day2 hosts was skipped for hosts with user managed networking enabled. We have removed this so that the validation will be conducted for all day 2 hosts now.