AGENT-227: Add host configuration support to ephemeral installer client (#4014)
* Ephemeral installer: refactor client command

In future we'll want to do different things with this, not just register
the Cluster and InfraEnv.

* Ephemeral installer: Split environment options for subcommands

* Add subcommands to ephemeral installer client

Change the command to register the Cluster and InfraEnv from:

    /agent-based-installer-register-cluster-and-infraenv

to:

    /usr/local/bin/agent-installer-client register

This allows us to add further subcommands to the same binary in future.

A symlink is provided for backward compatibility with the existing
agent-based installer ISOs. When invoked using the old name, the command
will retain the existing behaviour. This backward compatibility can be
removed once the agent-based installer is using the subcommand syntax.

* Implement host configuration framework for ephemeral installer

* Create library func to evaluate root device hints

* Set root device in ephemeral installer client

Read root device hints from (by default)
/etc/assisted/hostconfig/*/root-device-hints.yaml and use them to select
an install disk for each host and configure it through the assisted API.

* Add role configuration to ephemeral installer client

Read role files from (by default) /etc/assisted/hostconfig/*/role and
use them to set the role for each host through the assisted API.

* Record failures and log more useful information on the cause

(cherry picked from commit eba9079e61bb5022f788de9bc612fa53261b71e5)