MGMT-10780: Explicitly pass OCP version to Hypershift (#3965)
In order to control the version of the hosted cluster from the Prow job
definition, we are explicitly passing the `--release-image` to the
`hypershift create` command. This is leveraging the current design that
uses `ASSISTED_OPENSHIFT_INSTALL_RELEASE_IMAGE` to decide what is the
OCP version used in the spoke cluster in the ZTP flow.

Contributes-to: MGMT-10780

Co-authored-by: Mat Kowalski <mko@redhat.com>