AGENT-227: Don't exit when inventory not available
Before an agent's inventory has been discovered, we cannot match it to
any host configuration (which would require knowing the MAC addresses).
Therefore, treat it as equivalent to the host not having shown up yet.

Previously we exited with an error caused by trying to Unmarshal from
zero-length data, which results in the install failing when the
configuration command runs before inventory is available.