MGMTBUGSM-547: Add UnsyncedAgents error in ClusterDeployment and ACI
This PR adds a new "UnsyncedAgents" condition for marking clusters where
at least one of Agents has SpecSynced failure. Currently if this happens
we are showing a misleading error "UnapprovedAgents" coming from the
fact that data in the DB is not in sync with K8s CRs.

Cherry-picks: Bug-2076333
Cherry-picks: MGMTBUGSM-307
Closes: MGMTBUGSM-547