MGMT-10958: UpdateHostName - Validate that the hostname is not (#4029)
forbidden.

Presently when changing a hostname using UpdateHostName, no validation
is performed to ensure that that hostname is permitted.

This commit addresses that by validating that the hostname is not in
"localhost", "localhost.localdomain", "localhost4",
"localhost4.localdomain4", "localhost6.localdomain6"

This is to fix a bug that was introduced in MGMT-10722 and discovered during QA