NO-ISSUE: Update to Ginkgo 2
This patch updates the tests to use version 2 of Ginkgo. The only
relevant change is the replacement of the `.../ginkgo` package with
`.../ginkgo/v2/dsl/core`.

This update is necessary in order to use some of the metrics testing
utilities from the OCM SDK. Those utilities will be used in later
patches that change the metrics code. Without the update both version 1
and version 2 of Ginkgo are imported, and that generates problems,
specially when using the `ginkgo` tool.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>