MGMT-10844: Add virtual interface metrics for SaaS
https://issues.redhat.com/browse/MGMT-10844
As part of epic https://issues.redhat.com/browse/MGMT-10087
Virtual interface data per host and per cluster are available
in the SaaS environment, even if the feature is not enabled.

This PR records the virtual interfaces each host/cluster has and
exposes it as a metric. The data will help determine if this is a
viable feature we can enable in SaaS by allowing us to predict
performance/scalability and allowing us to see any patterns/common
interfaces.