MGMT-11026: Use image tag for metrics (#4078)
Currently the `assisted_installer_host_installation_phase_seconds_...`
metrics contain the agent version in the `discoveryAgentVersion` label, with
values like `v1.0.0-141`. This is copied directly from the
`discovery_agent_version` column of the `hosts` table. But with the
introduction of the agent upgrade feature this column will be changed to
store the full image reference instead of just the version number. That
means that the code that populates the metric labels needs to be changed
to detect if the `discovery_agent_version` is a full image reference,
and in that case extract the version from the tag of the image. That is
what this patch does.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>