MGMT-10848: Fix double agent unbind when deregistering a cluster (#3964)
* MGMT-10848: Fix double agent unbind when deregistering a cluster

When a cluster is removed in the kubeAPI the clusterdeployments
controller removes all the cluster references from all the agents. This
should cause the agent controller to unbind each of those agents.

The clusterdeployment controller also calls `DeregisterClusterInternal`
which also unbinds the hosts from the cluster.

This isn't a problem currently because both operations should be doing
the same thing, but it will become a problem when hosts should be moved
back to the infraEnv on unbind only if they are not associated with an
BMH, and the kubeAPI is being used. In this case the agent controller
will know if the host should be moved to the infraEnv or not, but the
unbind call from `DeregisterClusterInternal` would have to assume that
the hosts should _not_ be moved back to the infraEnv as it would also be
used in the SaaS REST API.

This difference in implementation would cause a race and needs to be
resolved.

This PR moves all the SaaS specific logic out of
`DeregisterClusterInternal` and into `V2DeregisterCluster`. This will
ensure that this logic is still run when it is needed, but avoid the
race with the controllers when using the kubeAPI.

https://issues.redhat.com/browse/MGMT-10848

* MGMT-10848: customizeHost never returns an error so remove the return value

* MGMT-10848: Don't require a host to have a cluster to get it by KubeKey

Originally this method just accessed the host, but the cluster
information was added to handle progress reporting.

This left the agent controller with no way to query the database for
hosts who are linked to a possibly deleted cluster.

This can happen when a ClusterDeployment is deleted while it still
had bound agents. The agents will be unbound asynchronously but were
failing because this method required the referenced cluster to still
exist.

* MGMT-10848: Fix lint error caused by previous changes

This line looked like nonsense even before, but I suppose something
about removing the err variable triggered a lint error

* MGMT-10848: Delete agents when clusterdeployment referenced by an infraEnv is deleted

This was previously being done by the internal API, but now that the
agent controller is handling unbinding the cluster deployment controller
needs to make the decision to delete the agent or unbind.