WIP: Discussion/proposal for validation refactor
The current host validation is performed in two phases, the "validate"
phase and the "print" phase. The response from then validate phase is
really just a simple string.

This leads to a problem, if a validation is more complex that a simple
set of statuses, it might be good to include a precise message on the
cause of the issue

This PR is a partial proposal on how we might achieve that