MGMT-8890: Fix wrong release in the error log (#3889)
This commit fixes a wrong release image path logged when extracting the
installer binary. This is for a scenario when the extraction happens
from the original OpenShift image but the error message tries to print
the user-provided mirror.

As in this path the mirror does not have to exist, we want to log the
original release image.