MGMT-9982: Alter ipxe-script endpoint to handle boot order (#4012)
When using iPXE would like to handle any boot order.  The options are
[network, hd], and [hd, network].
Currently we are only handling the second case, where we attempt to boot
from hd first.  If we fail, then we attempt to boot from network.
If network is first, we need to handle the case that hd is already
installed and host is rebooting, or installed.  In this case we would
like to avoid using the iPXE booting and boot from hd instead.