NO-ISSUE: Return bytes from `executeTemplate` template function
Currently the `executeTemplate` template function returns a string,
therefore assuming the result is text. This could make it difficult to
generate binary files. To make that more feasible this patch changes it
so that it returns an array of bytes. The conversion to UTF-8 encoded
strings will be performed by a new `toString` function.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>