MGMT-11131: Track hosts' registration times
This PR introduces a new host property RegisteredAt in order to track
all the registration attemps. This is because an already existing host
can e.g. reboot what would trigger a new registration attempt that
currently is not tracked as a property of the host even though it
triggers an event.

This is going to be helpful as there are scenarios when we want to know
if the host rebooted after the initial registration. With this change it
will be as simple as comparing timestamps of CreatedAt and RegisteredAt.

Contributes-to: MGMT-11131