MGMT-10570: check that are master machine found before referencing them (#4108)
An extermely rare scenario during development caused the service to crash
because there were no machine resources whne looking for the spoke
masters.

Even though this scenario is not expected to happen during regular flows
of installation, it is still good habit to check if the list of machine
has items in it before referencing them.