MGMT-11131: Use host's registration time as key for cache
It has been discovered that hosts can get different IP addresses when
they reboot post-discovery. Currently our connectivity validations use
only Host ID and InfraEnv ID as a key and never invalidate the entry.
This causes problems if the host changes its IP address, as the
validation can still rely on the old inventory.

Given that the cache exists inside the validator and is not connected
with the inventory and also the fact that unmarshalling the inventory is
an expensive operation, we decided to additionally use registration
timestamp as a key for the cache.

This limits the exposure to the described issue as with this change
whenever the host reboots it will use a new entry in the cache. Doing so
allows us to avoid redesigning the cache mechanism (the performance hit
for invalidating on every inventory update is too big) while solving the
observed issue.

Please note that if for any reason host changes its IP address without
rebooting or restarting the agent, the wrong cache may still be used.
This has been discussed internally and is at the moment accepted to be a
known issue.

Closes: MGMT-11131