Bug-2106496: Make sure that updateInventory function does not use the internal DB object. (#4136)
The updateInventory function in the host package is called with db
object parameter.  This may be a transactional object.  If it is, direct
use of the internal db object cause a new db connection to be requested
from the db connection pool.  If all the connections in the db
connection pool are in use, then this function will wait.  Since this
goroutine is holding a transaction, this may result in complete halt of
all database operations.
Changed to use only the db object received as parameter.