MGMT-10445: Add enhancement for automating returning an agent to the infra-env (#3849)
Without a BareMetalHost there is currently no way to automate the process
of unbinding a host from a cluster. When a user wants to reuse a host for a new
cluster they have to use the assisted-service unbind API, then manually boot
the host back into the discovery image.

To support automating this procedure, the assisted-service will configure nodes
in installed clusters to run an agent. When appropriate this agent will
reconfigure the host to boot into a discovery image, allowing the host to
be made available for installation without the user manually interacting with
a management system to reboot the host into the discovery image.

https://issues.redhat.com/browse/MGMT-10445