MGMT-10973 - kube-api imported clusters should be considered imported
As part of 564650feca372064314282d98d6fd9c6ee69bd2c we pretended
imported kube-api clusters are not imported, which is wrong, just
to disable DNS validations because those validations were causing
issues in the CAPI CI and we had to merge the PR quickly because
it was needed for a SaaS release.

This commit reverts that hack. We expect the capi CI to fail, but
at-least now we can fix it without hurrying to get it merged for
a release.