MGMT-11198: handle nmstate empty connection list
nmstate response to an invalid configuration has changed, as
described in: https://github.com/nmstate/nmstate/issues/1975

This becomes an issue when we use a newer nmstate versions, as done
in the work we intend to merge here: https://github.com/openshift/assisted-service/pull/3979

The kubeAPI subsystem test[1] makes sure we get the right response for each case.
It now passes both the the older (this PR) and a newer (https://github.com/openshift/assisted-service/pull/3979) versions
of nmstate.

[1] "deploy clusterDeployment and infraEnv and with an invalid NMState config YAML"