Bug-2106496: Do not perform refresh status on hosts after roles need to (#4137)
auto-assign during host registration
It is not needed since host registration is no interactive operation.

Do not perform select for update on InfraEnv during host registration.
This is not needed since the InfraEnv is not updated in this operation.
This avoids contension while many hosts register to the same InfraEnv.