MGMT-11174: Add IPXE boot control flag to infra env in kube API (#4151)
This is needed to support boot control in Kube API.
In Kube API, a new type IPXEScriptType is introduced which has 2
possible values - HostRedirect and Boot.
HostRedirect means that IPXE URL will cause serving the redirect script
to the host using IPXE, while Boot means that the boot script will be
served.