MGMT-11262: Parse `/etc/kubernetes/kubeconfig` in day-2 ignition to disable DNS validations (#4171)
When the ignition downloaded by the day-2 worker discovery indicates
that kubelet is connecting to the day-1 cluster using an IP address
rather than a domain name, we can safely disable DNS validations for
that worker in day-2.

This is the case in Hypershift.

The agent was modified to include /etc/kubernetes/kubeconfig in the
day-2 ignition to allow this, because previously it was omitted for size
concerns

This commit modifies the service to look at this file that is now
included by the agent and parse it to choose whether DNS validations
need to be enabled or disabled