NO-ISSUE: Don't log whole ignition for InfraEnv
This commit changes the log level from INFO to DEBUG for the content of
the ignition placed inside ISO generated for a specific InfraEnv. This
is so that when using a default log level we don't print the whole
content but only the message indicating that the generation has
happened.