MGMT-10734: Add more host reclaim states and transitions (#4139)
* MGMT-10734: Add TEST to skipper.env

This will allow specifying a package to test

* MGMT-10734: Allow reclaiming day2 hosts

* MGMT-10734: Add failure and rebooting reclaim transitions

This commit adds transitions for moving between the two reclaim states
(Reclaiming -> RebootingForReclaim) as well as a transition to
explicitly fail the reclaim process.

* MGMT-10734: Monitor reclaiming hosts for timeout

Ensure that the host monitor is checking hosts in either `Reclaiming` or
`ReclaimingRebooting` for timeouts on `Refresh` transitions.

This commit makes the previous installation timeout condition more
generally useful and reuses it here.

* MGMT-10734: add unbind-type transitions for reclaim

The reclaim transition should work the same as unbind in the states that are
not particularly interesting for the new work. This means that any time that
we don't need to reboot the host it should act exactly the same as unbind
already does.

https://issues.redhat.com/browse/MGMT-10734