MGMT-11362: Set api vip and not dnsname for day2 cluster (transform) for bm platform (#4191)
We see cases where user are not able to manage dns inside their org and
they are stuck with trying to add hosts. We know that in BM case there
is no need to set dns name if we can set api vip.
Changing tranform to day2 logic to set vip in case it can