MGMT-6537: Remove sqlite dependency / package
It was added during #2233 and was left unused.

It was supposed to be used during #2794 and #2831 but those PRs never
got merged and seem to have been deprioritized.

The go-sqlite3 dependency is huge, contains 100k LOC of C bindings that
give off annoying warnings (depending on your system's compiler
version), and bloats our vendor directory.

Also the usage of sqlite inside catalog images of operators has been
deprecated so I'm not even sure this is still the corect way to solve
this from a technical perspective