MGMT-10969: Ensure infraEnv name label is set on agents (#4198)
* Add GetInfraEnvInternal to the installer internal API interface

This will be needed when the agent controller attempts to label the
agent resource with the infraEnv name as it only currently has access to
the infraEnv ID.

https://issues.redhat.com/browse/MGMT-10969

* Ensure infraEnv label is set on agent resources

Previously the infraEnv name label was only set on agent resources when
the agent was originally created. If the label was changed or deleted
this would break processes that rely on it being present.

https://issues.redhat.com/browse/MGMT-10969