MGMT-8890: Handle registries.conf without any mirrors correctly
Currently the code assumes that existence of containers.conf file means
there are always registry mirrors configured. It has been observed that
the default file has no `[[registry]]` section but is still a correct
file. For such a scenario we want to proceed as the file did not exist
instead of failing.

Contributes-to: MGMT-8890